Py_REFCNT is now mandatory

+++ Include/lock_linux.h
@@ -2,7 +2,9 @@
 #define Py_LOCK_LINUX_H
 
 #include <linux/futex.h>
+#include <assert.h>
 #include <pthread.h>
+#include <string.h>
 #include <sys/mman.h>
 #include <sys/syscall.h>
 #include <sys/time.h>
@@ -11,14 +13,14 @@
 #include <x86intrin.h>
 
 typedef pthread_t threadid_t;
-typedef int primitivelock_t;
+typedef Py_ssize_t primitivelock_t;
 #define CURTHREAD_ID pthread_self
 #define THREAD_EQUAL(x, y) pthread_equal(x, y)
 #define ATOMIC_INC(x)  __sync_add_and_fetch(x, 1)
 #define ATOMIC_DEC(x)  __sync_sub_and_fetch(x, 1)
 #define ATOMIC_ADD(x, y)  __sync_add_and_fetch(x, y)
 
-Py_LOCAL_INLINE(int) syscall_futex(int *futex, int operation, int value) {
+Py_LOCAL_INLINE(int) syscall_futex(Py_ssize_t *futex, int operation, int value) {
     return syscall(SYS_futex, futex, operation, value, NULL, NULL, NULL);
 }
 
@@ -32,11 +34,11 @@ Py_LOCAL_INLINE(int) syscall_futex(int *futex, int operation, int value) {
 
 #define FUTEX_STATIC_INIT(description) { 0, description FUTEX_STATS_STATIC_INIT }
 
-Py_LOCAL_INLINE(void) futex_init_primitive(int *lock) {
+Py_LOCAL_INLINE(void) futex_init_primitive(Py_ssize_t *lock) {
 	*lock = 0;
 }
 
-Py_LOCAL_INLINE(void) futex_lock_primitive(int *lock) {
+Py_LOCAL_INLINE(void) futex_lock_primitive(Py_ssize_t *lock) {
 	int current = __sync_val_compare_and_swap(lock, 0, 1);
 	if (current == 0)
 		return;
@@ -48,7 +50,7 @@ Py_LOCAL_INLINE(void) futex_lock_primitive(int *lock) {
 	}
 }
 
-Py_LOCAL_INLINE(void) futex_unlock_primitive(int *lock) {
+Py_LOCAL_INLINE(void) futex_unlock_primitive(Py_ssize_t *lock) {
 	if (__sync_fetch_and_sub(lock, 1) != 1) {
 		*lock = 0;
 		futex_wake(lock, 1);
@@ -58,6 +60,5 @@ Py_LOCAL_INLINE(void) futex_unlock_primitive(int *lock) {
 PyAPI_FUNC(Py_ssize_t) Py_AtomicInc(Py_ssize_t* value);
 PyAPI_FUNC(Py_ssize_t) Py_AtomicDec(Py_ssize_t* value);
 PyAPI_FUNC(Py_ssize_t) Py_AtomicAdd(Py_ssize_t* to, Py_ssize_t value);
-
 
 #endif
